requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pymongo==3.13.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pytz==2024.2